+++ src/components/Table/TR/index.js
@@ -1,8 +1,14 @@
 import React from 'react';
+import classNames from 'classnames';
 
 const TR = (props) => {
+  let classes = classNames({
+    'oui-table-row--active': props.isActive,
+  });
   return (
-    <tr data-test-section={ props.testSection }>
+    <tr
+      className={ classes }
+      data-test-section={ props.testSection }>
       { props.children }
     </tr>
   );
@@ -11,6 +17,8 @@ const TR = (props) => {
 TR.propTypes = {
   /** Expects a `Table.TD` or `Table.TH` component */
   children: React.PropTypes.node,
+  /** If true, add active class */
+  isActive: React.PropTypes.bool,
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
 };